Move update module to cli_commands package
